requirements.txt
beautifulsoup4==4.14.3
chromadb==1.5.2
google-news-scraper==0.1.0
googlenewsdecoder==0.1.7
openai==2.24.0
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
streamlit==1.55.0
